Tests/LibWeb/Text/expected/input-file.txt

input1:
file1 (index iteration): text/plain: Contents for file1
file1 (for..of iteration): text/plain: Contents for file1
input2:
file1 (index iteration): text/plain: Contents for file1
file2 (index iteration): text/plain: Contents for file2
file3 (index iteration): text/plain: Contents for file3
file4 (index iteration): text/plain: Contents for file4
file1 (for..of iteration): text/plain: Contents for file1
file2 (for..of iteration): text/plain: Contents for file2
file3 (for..of iteration): text/plain: Contents for file3
file4 (for..of iteration): text/plain: Contents for file4
